task_feasibility/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.2=gpu_0
attrs=20.3.0=pyhd3eb1b0_0
blas=1.0=mkl
boto=2.49.0=py36_0
boto3=1.7.62=py36h28b3542_1
botocore=1.10.62=py36h28b3542_0
brotlipy=0.7.0=py36h8f6f2f9_1001
bz2file=0.98=py36_1
ca-certificates=2020.10.14=0
certifi=2020.6.20=py36_0
cffi=1.14.5=py36hc120d54_0
click=7.1.2=pyh9f0ad1d_0
cryptography=3.3=py36h3c74f83_0
cudatoolkit=11.0.221=h6bb024c_0
cudnn=6.0=0
dataclasses=0.7=py36_0
docutils=0.16=py36_1
fastrlock=0.5=py36he6710b0_0
filelock=3.0.12=pyh9f0ad1d_0
freetype=2.10.4=h5ab3b9f_0
gensim=3.8.0=py36h962f231_0
icu=68.1=h58526e2_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=3.7.0=py36h5fab9bb_0
importlib_metadata=1.5.0=py36_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2020.2=254
jmespath=0.10.0=py_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9b=habf39ab_1
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=7.3.0=hdf63c60_0
libgomp=9.3.0=h2828fa1_18
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.2=heb0550a_3
mkl=2019.4=243
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.0=py36hd6b4f25_0
more-itertools=8.7.0=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=py36hff7bd54_0
nltk=3.4.4=py_0
nodejs=14.15.1=h25f6087_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py_0
openssl=1.1.1h=h7b6447c_0
packaging=20.9=pyh44b312d_0
pillow=8.0.0=py36h9a89aac_0
pip=20.3.3=py36h06a4308_0
pluggy=0.13.1=py36h06a4308_0
progressbar2=3.37.1=py36h06a4308_0
py=1.10.0=pyhd3eb1b0_0
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py36h5fab9bb_3
pytest=6.2.3=py36h06a4308_2
pytest-runner=5.3.0=pyhd3eb1b0_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python-utils=2.5.6=py36h06a4308_0
python_abi=3.6=1_cp36m
pytorch=1.7.1=py3.6_cuda11.0.221_cudnn8.0.5_0
readline=8.1=h27cfd23_0
regex=2020.11.13=py36h8f6f2f9_1
requests=2.13.0=py36_0
s3transfer=0.1.13=py36_0
sacremoses=0.0.43=pyh9f0ad1d_0
scikit-learn=0.24.1=py36ha9443f7_0
scipy=1.5.2=py36h0b6359f_0
selenium=3.141.0=py36h8f6f2f9_1002
sentence-transformers=0.4.1=pyhd3deb0d_0
setuptools=52.0.0=py36h06a4308_0
six=1.15.0=pyh9f0ad1d_0
smart_open=1.9.0=py_0
sqlite=3.33.0=h62c20be_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tokenizers=0.9.4=py36h2bc52f9_1
toml=0.10.2=pyhd3eb1b0_0
torchvision=0.8.2=py36_cu110
tqdm=4.58.0=pyhd8ed1ab_0
transformers=4.3.2=pyhd8ed1ab_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.3=pyhd8ed1ab_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zipp=3.4.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
